revisi tipe data float -> double

+++ database/migrations/2020_06_24_060424_create_sales_table.php
@@ -22,7 +22,7 @@ class CreateSalesTable extends Migration
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
 
-            $table->float('total_payment');
+            $table->double('total_payment');
 
             $table->timestamps();
         });
